requirements.txt
googletrans==4.0.0rc1
gTTS==2.5.4
Kivy==2.3.1
moviepy==1.0.3
openai_whisper==20250625
pydub==0.25.1
pyrubberband==0.4.0
PyYAML==6.0.1
soundfile==0.12.1
